fix: run forge fmt to align with CI standards

+++ test/14_MoneyFlow.t.sol
@@ -14,14 +14,14 @@ contract MoneyFlowTest is Test {
 
     // Test Actors
     address public admin = address(0x1);
-    address public user = address(0x2);          // Payer (signs intent, gets cashback)
-    address public merchant = address(0x3);       // Merchant (receives payment)
+    address public user = address(0x2); // Payer (signs intent, gets cashback)
+    address public merchant = address(0x3); // Merchant (receives payment)
     address public user3 = address(0x4);
     address public keeper = address(0x5);
     address public treasury = address(0x6);
     address public validatorPool = address(0x7);
     address public dexBroker = address(0x8);
-    address public agent = address(0x777);        // Agent (executes transaction, pays from his wallet)
+    address public agent = address(0x777); // Agent (executes transaction, pays from his wallet)
 
     // Signer Configuration
     uint256 constant SIGNER_KEY = 0xA1;
@@ -32,9 +32,9 @@ contract MoneyFlowTest is Test {
     uint128 constant MAX_CLAIM_LIMIT = 5 ether;
     uint128 constant BURN_QUEUE_LIMIT = 10 ether;
     uint256 constant VESTING_BLOCKS = 7200;
-    uint256 constant FEE_BPS = 100;                     // 1%
-    uint256 constant BURN_SHARE_DIVISOR = 5;            // 20% of fee
-    
+    uint256 constant FEE_BPS = 100; // 1%
+    uint256 constant BURN_SHARE_DIVISOR = 5; // 20% of fee
+
     // Exact amount: 987,649,235 MON
     uint128 constant EXACT_AMOUNT = 987_649_235 ether;
 
@@ -51,14 +51,7 @@ contract MoneyFlowTest is Test {
         signer = vm.addr(SIGNER_KEY);
 
         vm.startPrank(admin);
-        axil = new AxilProtocolV1(
-            admin,
-            signer,
-            treasury,
-            validatorPool,
-            dexBroker,
-            keccak256("SALT")
-        );
+        axil = new AxilProtocolV1(admin, signer, treasury, validatorPool, dexBroker, keccak256("SALT"));
 
         axil.updateConfig(AxilProtocolV1.ConfigKey.MaxClaim, MAX_CLAIM_LIMIT, address(0));
         axil.updateConfig(AxilProtocolV1.ConfigKey.BurnLimit, BURN_QUEUE_LIMIT, address(0));
@@ -69,10 +62,10 @@ contract MoneyFlowTest is Test {
         vm.stopPrank();
 
         // Fund addresses
-        vm.deal(user, EXACT_AMOUNT * 2);        // User has 1,975,298,470 MON (but doesn't spend)
-        vm.deal(merchant, EXACT_AMOUNT * 2);    // 1,975,298,470 MON
-        vm.deal(user3, EXACT_AMOUNT * 2);       // 1,975,298,470 MON
-        vm.deal(agent, EXACT_AMOUNT * 2);       // Agent has 1,975,298,470 MON (HE spends)
+        vm.deal(user, EXACT_AMOUNT * 2); // User has 1,975,298,470 MON (but doesn't spend)
+        vm.deal(merchant, EXACT_AMOUNT * 2); // 1,975,298,470 MON
+        vm.deal(user3, EXACT_AMOUNT * 2); // 1,975,298,470 MON
+        vm.deal(agent, EXACT_AMOUNT * 2); // Agent has 1,975,298,470 MON (HE spends)
         vm.deal(keeper, 1000 ether);
         vm.deal(address(this), EXACT_AMOUNT * 2);
     }
@@ -99,85 +92,53 @@ contract MoneyFlowTest is Test {
         address _agent
     ) internal view returns (bytes memory) {
         bytes32 structHash = keccak256(
-
-abi.encode(
-                EXECUTE_TYPEHASH,
-                _merchant,
-                _user,
-                intentId,
-                amount,
-                deadline,
-                salt,
-                _agent
-            )
+            abi.encode(EXECUTE_TYPEHASH, _merchant, _user, intentId, amount, deadline, salt, _agent)
         );
         bytes32 finalHash = keccak256(abi.encodePacked("\x19\x01", _getDomainSeparator(), structHash));
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(SIGNER_KEY, finalHash);
         return abi.encodePacked(r, s, v);
     }
 
-    function _executePayment(
-        address _merchant,
-        address _user,
-        uint128 amount,
-        uint256 nonce
-    ) internal returns (bytes32) {
+    function _executePayment(address _merchant, address _user, uint128 amount, uint256 nonce)
+        internal
+        returns (bytes32)
+    {
         require(amount >= MIN_EXECUTION_AMOUNT, "Amount below minimum execution");
-        
-        bytes32 intentId = keccak256(abi.encodePacked(
-            block.timestamp,
-            _merchant,
-            _user,
-            amount,
-            nonce,
-            "payment"
-        ));
-        
+
+        bytes32 intentId = keccak256(abi.encodePacked(block.timestamp, _merchant, _user, amount, nonce, "payment"));
+
         uint256 deadline = block.timestamp + 1000;
-        uint128 salt = uint128(uint256(keccak256(abi.encodePacked(
-            block.timestamp,
-            _merchant,
-            _user,
-            amount,
-            nonce,
-            "salt"
-        ))));
-        
-        bytes memory signature = _createExecuteSignature(
-            _merchant,
-            _user,
-            intentId,
-            amount,
-            deadline,
-            salt,
-            address(this)
-        );
-        
+        uint128 salt =
+            uint128(uint256(keccak256(abi.encodePacked(block.timestamp, _merchant, _user, amount, nonce, "salt"))));
+
+        bytes memory signature =
+            _createExecuteSignature(_merchant, _user, intentId, amount, deadline, salt, address(this));
+
         vm.deal(address(this), amount);
         vm.prank(address(this));
         axil.execute{value: amount}(_merchant, _user, intentId, deadline, salt, signature);
-        
+
         return intentId;
     }
 
     function test_ExactAmountDistribution() public {
         uint128 transferAmount = EXACT_AMOUNT; // 987,649,235 MON
-        
+
         console.log("\n--- 5-WAY DISTRIBUTION TEST ---");
         console.log("Transfer Amount: 987,649,235 MON");
         console.log("Fee (1%):", uint256(transferAmount) / 100, "MON");
         console.log("----------------------------------------");
-        
+
         // Record pre-transaction state
-        uint256 userBalBefore = user.balance;           // User's balance (doesn't change)
-        uint256 agentBalBefore = agent.balance;         // Agent's balance (decreases)
-        uint256 merchantBalBefore = merchant.balance;   // Merchant's balance (increases)
-        
-        (uint128 userRewardsBefore, , , ) = axil.getRewardVault(user);
-        (uint128 agentRewardsBefore, , , ) = axil.getRewardVault(agent);
-        (uint128 validatorRewardsBefore, , , ) = axil.getRewardVault(validatorPool);
-        (uint128 dexRewardsBefore, , , ) = axil.getRewardVault(dexBroker);
-        
+        uint256 userBalBefore = user.balance; // User's balance (doesn't change)
+        uint256 agentBalBefore = agent.balance; // Agent's balance (decreases)
+        uint256 merchantBalBefore = merchant.balance; // Merchant's balance (increases)
+
+        (uint128 userRewardsBefore,,,) = axil.getRewardVault(user);
+        (uint128 agentRewardsBefore,,,) = axil.getRewardVault(agent);
+        (uint128 validatorRewardsBefore,,,) = axil.getRewardVault(validatorPool);
+        (uint128 dexRewardsBefore,,,) = axil.getRewardVault(dexBroker);
+
         uint256 burnedBefore = axil.totalBurned();
         uint256 failedBefore = axil.failedBurnQueue();
 
@@ -193,39 +154,32 @@ abi.encode(
         uint128 salt = 12345;
         uint128 amount = transferAmount;
         address agentAddr = agent;
-        
-        bytes memory signature = _createExecuteSignature(
-            merchantAddr,
-            userAddr,
-            intentId,
-            amount,
-            deadline,
-            salt,
-            agentAddr
-        );
-        
+
+        bytes memory signature =
+            _createExecuteSignature(merchantAddr, userAddr, intentId, amount, deadline, salt, agentAddr);
+
         vm.prank(agentAddr);
         axil.execute{value: amount}(merchantAddr, userAddr, intentId, deadline, salt, signature);
 
         // Calculate expected values - FIXED: split into separate lines to avoid stack too deep
 
-uint256 amountUint = uint256(transferAmount);
+        uint256 amountUint = uint256(transferAmount);
         uint256 feeWithBps = amountUint * FEE_BPS;
-        uint256 expectedFee = feeWithBps / 10000;                           // 9,876,492.35 MON
-        uint256 expectedBurn = expectedFee / BURN_SHARE_DIVISOR;            // 1,975,298.47 MON
-        uint256 expectedRewardShare = expectedFee - expectedBurn;           // 7,901,193.88 MON
-        uint256 expectedAgent = expectedRewardShare / 4;                    // 1,975,298.47 MON
-        uint256 expectedUser = expectedRewardShare / 4;                     // 1,975,298.47 MON
-        uint256 expectedValidator = expectedRewardShare / 4;                // 1,975,298.47 MON
-        uint256 expectedDex = expectedRewardShare - (expectedAgent * 3);    // 1,975,298.47 MON
-        uint256 expectedMerchant = amountUint - expectedFee;                // 977,772,742.65 MON
+        uint256 expectedFee = feeWithBps / 10000; // 9,876,492.35 MON
+        uint256 expectedBurn = expectedFee / BURN_SHARE_DIVISOR; // 1,975,298.47 MON
+        uint256 expectedRewardShare = expectedFee - expectedBurn; // 7,901,193.88 MON
+        uint256 expectedAgent = expectedRewardShare / 4; // 1,975,298.47 MON
+        uint256 expectedUser = expectedRewardShare / 4; // 1,975,298.47 MON
+        uint256 expectedValidator = expectedRewardShare / 4; // 1,975,298.47 MON
+        uint256 expectedDex = expectedRewardShare - (expectedAgent * 3); // 1,975,298.47 MON
+        uint256 expectedMerchant = amountUint - expectedFee; // 977,772,742.65 MON
 
         // Record post-transaction state
-        (uint128 userRewardsAfter, , , ) = axil.getRewardVault(user);
-        (uint128 agentRewardsAfter, , , ) = axil.getRewardVault(agent);
-        (uint128 validatorRewardsAfter, , , ) = axil.getRewardVault(validatorPool);
-        (uint128 dexRewardsAfter, , , ) = axil.getRewardVault(dexBroker);
-        
+        (uint128 userRewardsAfter,,,) = axil.getRewardVault(user);
+        (uint128 agentRewardsAfter,,,) = axil.getRewardVault(agent);
+        (uint128 validatorRewardsAfter,,,) = axil.getRewardVault(validatorPool);
+        (uint128 dexRewardsAfter,,,) = axil.getRewardVault(dexBroker);
+
         uint256 burnedAfter = axil.totalBurned();
         uint256 failedAfter = axil.failedBurnQueue();
 
@@ -237,53 +191,29 @@ uint256 amountUint = uint256(transferAmount);
         console.log("----------------------------------------");
 
         // VERIFICATION 1: USER'S BALANCE DOES NOT CHANGE (he only signs)
-        assertEq(
-            user.balance,
-            userBalBefore,
-            "User's balance should not change (he only signs)"
-        );
+        assertEq(user.balance, userBalBefore, "User's balance should not change (he only signs)");
         console.log(" User's balance unchanged (only signs)");
 
         // VERIFICATION 2: AGENT spends exact amount
-        assertEq(
-            agent.balance,
-            agentBalBefore - transferAmount,
-            "Agent should spend exact amount"
-        );
+        assertEq(agent.balance, agentBalBefore - transferAmount, "Agent should spend exact amount");
         console.log(" Agent spent:", transferAmount, "MON");
 
         // VERIFICATION 3: Merchant receives 99% instantly
-        assertEq(
-            merchant.balance - merchantBalBefore,
-            expectedMerchant,
-            "Merchant should receive 99% instantly"
-        );
+        assertEq(merchant.balance - merchantBalBefore, expectedMerchant, "Merchant should receive 99% instantly");
         console.log("\nINSTANT TRANSFERS:");
         console.log("  Merchant (99%):", expectedMerchant, "MON");
 
         // VERIFICATION 4: Burn happens instantly (0.2%)
-        assertEq(
-            burnedAfter - burnedBefore,
-            expectedBurn,
-            "Burn should receive 0.2% instantly"
-        );
+        assertEq(burnedAfter - burnedBefore, expectedBurn, "Burn should receive 0.2% instantly");
         console.log("  Burn (0.2%):", expectedBurn, "MON");
 
         // VERIFICATION 5: Agent receives 0.2% pending rewards
-        assertEq(
-            agentRewardsAfter - agentRewardsBefore,
-            expectedAgent,
-            "Agent should receive 0.2% pending"
-        );
+        assertEq(agentRewardsAfter - agentRewardsBefore, expectedAgent, "Agent should receive 0.2% pending");
         console.log("\nPENDING REWARDS (vesting 7200 blocks):");
         console.log("  Agent (0.2%):", expectedAgent, "MON");
 
         // VERIFICATION 6: User receives 0.2% pending rewards (cashback)
-        assertEq(
-            userRewardsAfter - userRewardsBefore,
-            expectedUser,
-            "User should receive 0.2% cashback pending"
-        );
+        assertEq(userRewardsAfter - userRewardsBefore, expectedUser, "User should receive 0.2% cashback pending");
         console.log("  User (0.2% cashback):", expectedUser, "MON");
 
         // VERIFICATION 7: Validator Pool receives 0.2% pending rewards
@@ -294,12 +224,8 @@ uint256 amountUint = uint256(transferAmount);
         );
         console.log("  Validator Pool (0.2%):", expectedValidator, "MON");
 
-// VERIFICATION 8: DEX Broker receives 0.2% pending rewards
-        assertEq(
-            dexRewardsAfter - dexRewardsBefore,
-            expectedDex,
-            "DEX Broker should receive 0.2% pending"
-        );
+        // VERIFICATION 8: DEX Broker receives 0.2% pending rewards
+        assertEq(dexRewardsAfter - dexRewardsBefore, expectedDex, "DEX Broker should receive 0.2% pending");
         console.log("  DEX Broker (0.2%):", expectedDex, "MON");
 
         // VERIFICATION 9: No failed burns
@@ -317,7 +243,7 @@ uint256 amountUint = uint256(transferAmount);
         vm.prank(admin);
         axil.updateConfig(AxilProtocolV1.ConfigKey.BurnLimit, 0.1 ether, address(0));
 
-        for (uint i = 0; i < 5; i++) {
+        for (uint256 i = 0; i < 5; i++) {
             _executePayment(merchant, user, 100 ether, i + 1000);
         }
 
@@ -329,7 +255,7 @@ uint256 amountUint = uint256(transferAmount);
         vm.prank(admin);
         axil.updateConfig(AxilProtocolV1.ConfigKey.BurnCooldown, 0, address(0));
 
-        for (uint i = 0; i < 5; i++) {
+        for (uint256 i = 0; i < 5; i++) {
             _executePayment(merchant, user, 100 ether, i + 2000);
         }
 
@@ -358,7 +284,7 @@ uint256 amountUint = uint256(transferAmount);
 
         address[] memory accounts = new address[](1);
         accounts[0] = address(failingContract);
-        
+
         bytes32[] memory intents = new bytes32[](1);
         intents[0] = keccak256("claim-fail");
 
@@ -372,8 +298,8 @@ uint256 amountUint = uint256(transferAmount);
     function test_DistributionEdgeCases() public {
         _executePayment(merchant, user, MIN_EXECUTION_AMOUNT, 4000);
         _executePayment(merchant, user, 100 ether, 4001);
-        
-        (uint256 totalIntents, , , , , ) = axil.getSystemStats();
+
+        (uint256 totalIntents,,,,,) = axil.getSystemStats();
         assertTrue(totalIntents >= 2, "Both payments should be processed");
         console.log("Distribution edge cases covered");
     }
@@ -381,13 +307,13 @@ uint256 amountUint = uint256(transferAmount);
     function test_VestingAndRelease() public {
         _executePayment(merchant, user, 1000 ether, 5000);
 
-        (uint128 totalAmount, uint64 releaseBlock, , ) = axil.getRewardVault(address(this));
-        
+        (uint128 totalAmount, uint64 releaseBlock,,) = axil.getRewardVault(address(this));
+
         console.log("Agent rewards:", totalAmount);
         console.log("Current block:", block.number);
         console.log("Release block:", releaseBlock);
         console.log("Vesting blocks:", VESTING_BLOCKS);
-        
+
         assertTrue(totalAmount > 0, "Agent rewards should be accrued");
         assertTrue(releaseBlock > block.number, "Release block should be in future");
 
@@ -396,11 +322,11 @@ uint256 amountUint = uint256(transferAmount);
         vm.prank(address(this));
         axil.claimRewards(totalAmount, keccak256("vested-claim"));
 
-        (uint128 remaining, , , ) = axil.getRewardVault(address(this));
+        (uint128 remaining,,,) = axil.getRewardVault(address(this));
         assertEq(remaining, 0, "All rewards should be claimed after vesting");
     }
 
-function test_ComplexMultiUserFlow() public {
+    function test_ComplexMultiUserFlow() public {
         _executePayment(merchant, user, 500 ether, 6000);
         _executePayment(user, user3, 300 ether, 6001);
         _executePayment(user3, merchant, 200 ether, 6002);
@@ -416,10 +342,10 @@ function test_ComplexMultiUserFlow() public {
         vm.prank(keeper);
         axil.autoBatchClaim(accounts, intents, 1_000_000);
 
-        (uint256 totalIntents, uint256 totalValue, , , , ) = axil.getSystemStats();
+        (uint256 totalIntents, uint256 totalValue,,,,) = axil.getSystemStats();
         assertTrue(totalIntents >= 3, "Should have at least 3 intents");
         assertTrue(totalValue >= 1000 ether, "Should have processed significant value");
-        
+
         console.log("Complex flow completed");
     }
 
@@ -427,7 +353,7 @@ function test_ComplexMultiUserFlow() public {
         vm.prank(admin);
         axil.updateConfig(AxilProtocolV1.ConfigKey.BurnCooldown, 60, address(0));
 
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             _executePayment(merchant, user, 100 ether, 7000 + i);
         }
 
@@ -462,6 +388,7 @@ contract FailingReceiver {
     receive() external payable {
         revert("This contract does not accept ETH");
     }
+
     fallback() external payable {
         revert("This contract does not accept ETH");
     }
